Installer: Drop explicit pytest
pytest-cov depends on the appropriate version of pytest
and coverage depending on the Python implementation and version.
Explicitly listing pytest before pytest-cov confuses setuptools
and/or pip as they process the requirements in order, so
the pytest-cov constraints are too late to be effective and the
result is pluggy.PluginValidationError due to version conflicts.

Closes https://github.com/hotoffthehamster/click-alias/issues/2